+++ Api.lua
@@ -126,3 +126,10 @@ function AshBlzSkinApi.PowerBarVisible()
         return DB.Appearance.PowerBar.Visibility == Visibility.SHOW_ALWAYS
     end)
 end
+
+__Static__() __AutoCache__()
+function AshBlzSkinApi.DisplayOnlyDispellableDebuffs()
+    return AshBlzSkinApi.OnConfigChanged():Map(function()
+        return DB.Appearance.DisplayOnlyDispellableDebuffs
+    end)
+end