+++ Api_BCC.lua
@@ -152,10 +152,6 @@ __Static__() __AutoCache__()
 function AshBlzSkinApi.UnitDebuffCanDispell()
     local class = UnitClassBase("player")
     return Wow.UnitAura():Map(function(unit)
-            -- 在副本内才工作
-            local inInstance, instanceType = IsInInstance()
-            if not inInstance or instanceType == "none" then return false end
-
             local canDispell, canDispellType
             local index = 1
             local name = ""
@@ -175,5 +171,13 @@ __Static__() __AutoCache__()
 function AshBlzSkinApi.UnitDebuffCanDispellColor()
     return AshBlzSkinApi.UnitDebuffCanDispell():Map(function(_, dType)
         return GetDispellDebuffColor(dType or "")
+    end)
+end
+
+-- Debuff 过滤类型
+__Static__() __AutoCache__()
+function AshBlzSkinApi.DebuffAuraFilter()
+    return Wow.FromUnitEvent(Wow.FromEvent("AshToAsh_Blizzard_Skin_Config_Changed"):Map("=> 'any'")):Map(function(unit)
+        return UnitCanAttack("player", unit) and "PLAYER|HARMFUL" or (DB.Appearance.DisplayOnlyDispellableDebuffs and "RAID" or "HARMFUL")
     end)
 end